// app/src/app/models/tag-information.ts
/***************************************************************************
*
*  Copyright 2021 Google Inc.
*
*  Licensed under the Apache License, Version 2.0 (the "License");
*  you may not use this file except in compliance with the License.
*  You may obtain a copy of the License at
*
*      https://www.apache.org/licenses/LICENSE-2.0
*
*  Unless required by applicable law or agreed to in writing, software
*  distributed under the License is distributed on an "AS IS" BASIS,
*  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
*  See the License for the specific language governing permissions and
*  limitations under the License.
*
*  Note that these code samples being shared are not official Google
*  products and are not formally supported.
*
***************************************************************************/

export interface TagInformation {
    id: string;
    page: string;
    tagType: string;
    accountId: string;
    gTag: string;
    networkCall: string;
    floodlightId: string;
    floodlightActivityTag: string;
    floodlightActivityGroup: string;
    floodlightSalesOrder: string;
    floodlightUVariables: string;
    warnings: string;
    errors: string;
}